Struct memory_addr::VirtAddr

pub struct VirtAddr(/* private fields */);

A virtual memory address.

It’s a wrapper type around an usize.

Implementations

impl VirtAddr

pub const fn from(addr: usize) -> Self

Converts an usize to a virtual address.

pub const fn as_usize(self) -> usize

Converts the address to an usize.

pub const fn as_ptr(self) -> *const u8

Converts the virtual address to a raw pointer.

pub const fn as_mut_ptr(self) -> *mut u8

Converts the virtual address to a mutable raw pointer.

pub fn align_down<U>(self, align: U) -> Self
where U: Into<usize>,

Aligns the address downwards to the given alignment.

See the align_down function for more information.

pub fn align_up<U>(self, align: U) -> Self
where U: Into<usize>,

Aligns the address upwards to the given alignment.

See the align_up function for more information.

pub fn align_offset<U>(self, align: U) -> usize
where U: Into<usize>,

Returns the offset of the address within the given alignment.

See the align_offset function for more information.

pub fn is_aligned<U>(self, align: U) -> bool
where U: Into<usize>,

Checks whether the address has the demanded alignment.

See the is_aligned function for more information.

pub fn align_down_4k(self) -> Self

Aligns the address downwards to 4096 (bytes).

pub fn align_up_4k(self) -> Self

Aligns the address upwards to 4096 (bytes).

pub fn align_offset_4k(self) -> usize

Returns the offset of the address within a 4K-sized page.

pub fn is_aligned_4k(self) -> bool

Checks whether the address is 4K-aligned.